Update Add Stock Modal * Move functions to different components
